requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=23.0
django-environ>=0.10.0
pytest>=7.0
django-redis==5.4.0